// src/types/graphql.d.ts
export type Exact<T extends { [key: string]: unknown }> = {
  [K in keyof T]: T[K];
};
// ...
export type Scalars = {
  ID: { input: string; output: string };
  String: { input: string; output: string };
  Boolean: { input: boolean; output: boolean };
  Int: { input: number; output: number };
  Float: { input: number; output: number };
  bigint: { input: bigint; output: bigint };
  jsonb: { input: unknown; output: unknown };
  numeric: { input: bigint | number; output: bigint | number };
};
// ...
export type GetStandingsQueryVariables = Exact<{
  season: Scalars['Int']['input'];
}>;

export type GetStandingsQuery = {
  __typename?: 'query_root';
  events: Array<{
    __typename?: 'events';
    round_number?: number | null;
    name?: string | null;
  }>;
  drivers: Array<{
    __typename?: 'drivers';
    abbreviation?: string | null;
    full_name?: string | null;
    latest_constructor: Array<{
      __typename?: 'driver_sessions';
      constructor?: {
        __typename?: 'constructors';
        name?: string | null;
        color?: string | null;
      } | null;
    }>;
    driver_standings: Array<{
      __typename?: 'driver_standings';
      round?: number | null;
      points?: bigint | number | null;
      position?: number | null;
    }>;
  }>;
  constructors: Array<{
    __typename?: 'constructors';
    name?: string | null;
    color?: string | null;
    constructor_standings: Array<{
      __typename?: 'constructor_standings';
      round?: number | null;
      points?: bigint | number | null;
      position?: number | null;
    }>;
  }>;
};